Added functionality to the createPrintMediaCollectableView.fxml by having functional user input fields. Added the ability to create a PrintMedia object. Added functionality to the collectionManagerMenuView.fxml 'create' button and 'Select Category' ComboBox. Added simple toString method to PrintMedia class. Also minor changes to the Model classes.

+++ src/models/Figure.java
@@ -20,7 +20,7 @@ public class Figure extends CollectibleCollection{
      * @param origin - Name of the series the character is from
      * @param scale - scale category (1/12, 1/10, 1/8, 1/7, 1/6, 1/4, 1/3, 1/2, 1/1, n/a)
      */
-    public Figure(String itemName, double price, String itemCondition, String collectibleCategory, String companyName, String characterName, String origin, String scale) {
+    public Figure(String itemName, int price, String itemCondition, String collectibleCategory, String companyName, String characterName, String origin, String scale) {
         super(itemName, price, itemCondition, collectibleCategory);
         setCompanyName(companyName);
         setCharacterName(characterName);
